+++ src/oracle-database/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+2.0.0
+++++++
+* Added new resources , Network anchors, Resource anchors, BaseDb , ExaScale
+
 1.0.0
 ++++++
 * Initial release.